0.14.11-0.19.17

+++ api/task.ts
@@ -12,13 +12,24 @@ const app = new Hono<jwtEnv>();
 const tasksMap = new Map<string, Map<string, ITask>>();
 const syncTasks = async (username: string, collectionTask: Collection<ITask>) => {
     let serverTasks = tasksMap.get(username);
-    if (!serverTasks) {
-        serverTasks = new Map<string, ITask>();
-        tasksMap.set(username, serverTasks);
-        for await (const task of collectionTask.find())
-            serverTasks.set(task.word, task);
-    }
+    if (serverTasks) return serverTasks;
+    serverTasks = new Map<string, ITask>();
+    tasksMap.set(username, serverTasks);
+    for await (const task of collectionTask.find())
+        serverTasks.set(task.word, task);
     return serverTasks;
+}
+const upsertTask = async (ctask: ITask, serverTasks: Map<string, ITask>, collectionTask: Collection<ITask>) => {
+    delete ctask._id;
+    const stask = serverTasks.get(ctask.word);
+    if (!stask) {
+        serverTasks.set(ctask.word, ctask);
+        await collectionTask.insertOne(ctask);
+    } else if (ctask.last > stask.last) {
+        serverTasks.set(ctask.word, ctask);
+        const $set = { last: new Int32(ctask.last), next: new Int32(ctask.next), level: new Int32(ctask.level) } as any;
+        await collectionTask.updateOne({ word: ctask.word }, { $set });
+    }
 }
 
 app.post(user, auth, async (c) => {
@@ -26,32 +37,29 @@ app.post(user, auth, async (c) => {
     const collectionTask = getCollectionTask(username);
     const serverTasks = await syncTasks(username, collectionTask);
     const clientTasks: Array<ITask> = await c.req.json();
-    for (const ctask of clientTasks) {
-        delete ctask._id;
-        if (serverTasks.has(ctask.word)) {
-            const stask = serverTasks.get(ctask.word)!;
-            if (ctask.last > stask.last) {
-                serverTasks.set(ctask.word, ctask);
-                const $set = { last: new Int32(ctask.last), next: new Int32(ctask.next), level: new Int32(ctask.level) } as any;
-                await collectionTask.updateOne({ word: ctask.word }, { $set });
-            }
-        } else {
-            serverTasks.set(ctask.word, ctask);
-            await collectionTask.insertOne(ctask);
-        }
-    }
+    if (!clientTasks || !Array.isArray(clientTasks)) return emptyResponse(STATUS_CODE.BadRequest);
+    for (const ctask of clientTasks) upsertTask(ctask, serverTasks, collectionTask);
     console.log(`API task POST ${username} with tasks ${clientTasks.length}, return ${serverTasks.size}.`);
-    return jsonResponse(Array.from(serverTasks.values().map(task=>(delete task._id, task))));
+    return jsonResponse(Array.from(serverTasks.values().map(task => (delete task._id, task))));
 }).delete(user, auth, async (c) => {
     const username = c.get('username');
-    const words: Array<string> = await c.req.json();
     const collectionTask = getCollectionTask(username);
     const serverTasks = await syncTasks(username, collectionTask);
-    if (!Array.isArray(words)) return emptyResponse(STATUS_CODE.BadRequest);
+    const words: Array<string> = await c.req.json();
+    if (!words || !Array.isArray(words)) return emptyResponse(STATUS_CODE.BadRequest);
     for (const word of words) serverTasks.delete(word);
     const deleteResult = await collectionTask.deleteMany({ word: { $in: words } });
     console.log(`API task DELETE ${username} with tasks ${deleteResult.deletedCount}.`);
     return jsonResponse(deleteResult);
+}).put(user, auth, async (c) => {
+    const username = c.get('username');
+    const collectionTask = getCollectionTask(username);
+    const serverTasks = await syncTasks(username, collectionTask);
+    const ctask: ITask = await c.req.json();
+    if (!ctask) return emptyResponse(STATUS_CODE.BadRequest);
+    upsertTask(ctask, serverTasks, collectionTask);
+    console.log(`API task PUT ${username} with task ${ctask.word}.`);
+    return emptyResponse();
 })
 
 export default app;
